requirements.txt
black==24.4.2
Django==4.2.2
djangorestframework==3.15.2
flake8==7.1.0
ipython==8.25.0
isort==5.13.2
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-filter==24.2
djangorestframework-simplejwt==5.3.1
coverage
drf-yasg
celery==5.4.0
redis==5.0.6
django-celery-beat==2.6.0
eventlet==0.36.1
requests
django-cors-headers
